\name{ratingTV}
 \alias{ratingTV}
 \title{Rating TV in Colombia}
 \description{
  An application for the study of Colombian rating television data for a given day and the resulting
  typology of television viewers.
 }
 \usage{data(ratingTV)}
 \format{Object of class \code{list} with: \code{$tab} a \code{data.frame} of 1306 rows and 90 columns; \code{$ilus} a \code{data.frame} of 1306 rows and 4 colums of sociodemographic variables; and a \code{vector} \code{weights} with the 1306 weights of the rows.
 }
 \source{The Data were provided by professor Juan de la Rosa Ramos}
 \references{Corrales, M.L., & Pardo, C.E. (2015). Analisis de datos longitudinales cualitativos con analisis de correspondencias y clasificacion. Comunicaciones en Estadistica, 8(1), 11-32.
\url{http://dx.doi.org/10.15332/s2027-3355.2015.0001.01} 
}
 \keyword{datasets}
