\name{zonas}
 \alias{zonas}
 \title{Zonas}
 \description{
  An application for the study of Colombian rating television data for a given day and the resulting
  typology of television viewers.
 }
 \usage{data(zonas)}
 \format{Object whit class \code{data.frame} of 1306 rows and 90 columns.}
 \source{The Data were provided by professor Juan de la Rosa Ramos}
\references{Corrales, M.L., & Pardo, C.E. (2015). Analisis de datos longitudinales cualitativos con analisis de correspondencias y clasificacion. Comunicaciones en Estadistica, 8(1), 11-32.
\url{http://dx.doi.org/10.15332/s2027-3355.2015.0001.01}
}
 \keyword{datasets}
